refactor: available calendar, removed OLCalendar
